requirements.txt
annotated-types==0.6.0
anyio==3.7.1
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
humanfriendly==10.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pip==23.2.1
platformdirs==4.0.0
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
setuptools==65.5.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
